+++ psutil/arch/sunos/disk.c
@@ -10,8 +10,6 @@
 #include <sys/mnttab.h>
 #include <kstat.h>
 
-#include "../../arch/all/init.h"
-
 
 PyObject *
 psutil_disk_io_counters(PyObject *self, PyObject *args) {