+++ HISTORY.rst
@@ -1,5 +1,16 @@
 *Bug tracker at https://github.com/giampaolo/psutil/issues*
 
+
+7.1.1
+=====
+
+XXXX-XX-XX
+
+**Bug fixes**
+
+- 2641_, [SunOS]: cannot compile psutil from sources due to missing C include.
+- 2357_, [SunOS]: `Process.cmdline()`_ does not handle spaces properly. (patch by Ben Raz)
+
 7.1.0
 =====